feat: update fodler structure and rename apollon2 as apollonEditor
